// src/app/models/registration.model.ts
export class Registration {

    constructor() {
        this.id = 0;
        this.firstname = "";
        this.lastname = "";
        this.fullname = "";
        this.email = "";
        this.password = ""
        this.address = ""
        this.phone = "";

    }
    id: number;
    firstname: string;
    lastname: string;
    fullname: string
    email: string;
    phone: string;
    password: string;
    address: string;

}